+++ R/joyn-merge.R
@@ -163,7 +163,7 @@ joyn <- function(x,
                  update_values    = FALSE,
                  update_NAs       = update_values,
                  reportvar        = getOption("joyn.reportvar"),
-                 reporttype       = c("character", "numeric"),
+                 reporttype       = c("factor", "character", "numeric"),
                  roll             = NULL,
                  keep_common_vars = FALSE,
                  sort             = FALSE,
@@ -322,40 +322,19 @@ joyn <- function(x,
   #             include report variable   ---------
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-  yvars_w <- c(y_vars_to_keep, ".yreport") # working yvars
-  x <- x |>
-    ftransform(.xreport = 1)
-  y <- y |>
-    ftransform(.yreport = 2)
-
-  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  #                   Actual merge   ---------
-  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-  # simple merge
-  i.yvars <- paste0("i.", yvars_w)
-
-  # keep relevant variables in y
-  y <- y |> fselect(
-    by, yvars_w
-  )
-
-  # Perform workhorse join
-  x <- joyn_workhorse(
-    x          = x,
-    y          = y,
-    by         = by,
-    match_type = match_type,
-    suffixes   = suffixes,
-    sort       = sort
-  )
+  yvars_w <- y_vars_to_keep # working yvars
+  # yvars_w <- c(y_vars_to_keep, ".yreport") # working yvars
+  # x <- x |>
+  #   ftransform(.xreport = 1)
+  # y <- y |>
+  #   ftransform(.yreport = 2)
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   #                   Report variable   ---------
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
   # replace NAs in report vars
-  setnafill(x, fill = 0, cols = c(".xreport", ".yreport"))
+  # setnafill(x, fill = 0, cols = c(".xreport", ".yreport"))
 
   # report variable
   dropreport <- FALSE
@@ -375,45 +354,39 @@ joyn <- function(x,
       check_names <- make.names(check_names, unique = TRUE)
       nrv         <- setdiff(check_names, xnames)
 
-      store_joyn_msg(info = "reportvar {.strongVar {reportvar}}  is already part of the resulting table. It will be changed to {.strongVar {nrv}}")
+      store_joyn_msg(info = "reportvar {.strongVar {reportvar}}  is
+                     already part of the resulting table. It will be
+                     changed to {.strongVar {nrv}}")
 
       reportvar <- nrv
     }
   }
 
 
-  # report variable
-  collapse::settransform(x, use_util_report = .xreport + .yreport)
-  # Can this be done more efficiently with collapse?
-  data.table::setnames(x, "use_util_report", reportvar)
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  #                   Actual merge   ---------
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+  # simple merge
+  i.yvars <- paste0("i.", yvars_w)
+
+  # keep relevant variables in y
+  y <- y |>
+    fselect(by, yvars_w)
+
+  # Perform workhorse join
+  jn <- joyn_workhorse(
+    x          = x,
+    y          = y,
+    by         = by,
+    suffixes   = suffixes,
+    sort       = sort,
+    reportvar  = reportvar
+  )
 
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  #                   Filter rows - `keep`   ---------
-  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-  ## rows to keep -----
-  if (keep  %in% c("master", "left") ) {
-
-    x <- x |>
-      fsubset(get(reportvar)  != 2)
-
-  } else if (keep  %in% c("using", "right") ) {
-
-    x <- x |>
-      fsubset(get(reportvar)  != 1)
-
-  } else if (keep  == "inner") {
-
-    x <- x |>
-      fsubset(get(reportvar)  >= 3)
-  } else if (keep == "anti") {
-    x <- x |>
-      fsubset(get(reportvar) == 1)
-  }
-
-  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  #                   Update x   ---------
+  #                   Update jn   ---------
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   var_use <- NULL
   if (isTRUE(update_NAs) || isTRUE(update_values)) {
@@ -422,7 +395,7 @@ joyn <- function(x,
 
   if (isTRUE(update_NAs || update_values) & length(var_use) > 0 ) {
 
-      x <- update_na_values(dt           = x,
+    jn <- update_na_values(dt           = jn,
                             var          = var_use,
                             reportvar    = reportvar,
                             suffixes     = suffixes,
@@ -430,6 +403,30 @@ joyn <- function(x,
                             rep_values   = update_values
       )
 
+  }
+
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  #                   Filter rows - `keep`   ---------
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+  ## rows to keep -----
+  if (keep  %in% c("master", "left") ) {
+
+    jn <- jn |>
+      fsubset(get(reportvar)  != 2)
+
+  } else if (keep  %in% c("using", "right") ) {
+
+    jn <- jn |>
+      fsubset(get(reportvar)  != 1)
+
+  } else if (keep  == "inner") {
+
+    jn <- jn |>
+      fsubset(get(reportvar)  >= 3)
+  } else if (keep == "anti") {
+    jn <- jn |>
+      fsubset(get(reportvar) == 1)
   }
 
 
@@ -441,15 +438,15 @@ joyn <- function(x,
       gsub("\\.", "\\\\.", x = _) |>
       paste0("$")
 
-    varx <- grep(patterns[1], names(x), value = TRUE)
-    vary <- grep(patterns[2], names(x), value = TRUE)
+    varx <- grep(patterns[1], names(jn), value = TRUE)
+    vary <- grep(patterns[2], names(jn), value = TRUE)
 
     # delete Y vars with suffix
-    collapse::get_vars(x, vary) <- NULL
+    collapse::get_vars(jn, vary) <- NULL
 
     # remove suffixes
     nsvar <- gsub(patterns[1], "", varx)
-    data.table::setnames(x, varx, nsvar)
+    data.table::setnames(jn, varx, nsvar)
 
   }
 
@@ -458,20 +455,43 @@ joyn <- function(x,
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
   ## cleaning temporary report variables ----
-  collapse::settransform(x,
-                         .xreport = NULL,
-                         .yreport = NULL)
+  # collapse::settransform(jn,
+  #                        .xreport = NULL,
+  #                        .yreport = NULL)
 
   # Rename by variables -----
   ## in output
-  if (any(grepl(pattern = "keyby", x = names(x)))) {
-    data.table::setnames(x,
-                         old = names(x)[grepl(pattern = "keyby",
-                                              x       = names(x))],
+  if (any(grepl("keyby",names(jn)))) {
+    data.table::setnames(jn,
+                         old = names(jn)[grepl("keyby", names(jn))],
                          new = xbynames)
   }
 
+
+  # Report variable ----------
+  # no matching obs
+  if (all(jn[[reportvar]] %in% c(1, 2)) && !keep == "anti") {
+
+    store_joyn_msg(warn = " you have no matching obs. Make sure argument
+                           `by` is correct. Right now, `joyn` is joining by
+                           {.strongVar {by}}")
+
+  }
+
   ## convert to characters if chosen -------
+
+  if (reporttype == "factor") {
+
+    get_vars(jn, reportvar) <-  factor(jn[[reportvar]],
+                                      levels = 1:6,
+                                      labels = c("x",
+                                                 "y",
+                                                 "x & y",
+                                                 "NA updated",
+                                                 "value updated",
+                                                 "not updated"))
+  }
+
   if (reporttype == "character") {
 
     rvar_to_chr <- \(x) {
@@ -484,29 +504,27 @@ joyn <- function(x,
                         default = "conflict")
     }
 
-    settransformv(x, reportvar, rvar_to_chr)
+    settransformv(jn, reportvar, rvar_to_chr)
 
   }
 
-  # no matching obs
-  if ((all(x[[reportvar]] %in% c("x", "y")) ||
-      all(x[[reportvar]] %in% c(1, 2))) &&
-      !keep == "anti") {
-
-    store_joyn_msg(warn = " you have no matching obs. Make sure argument
-                           `by` is correct. Right now, `joyn` is joining by
-                           {.strongVar {by}}")
-
-  }
 
   ## Display results------
   # freq table
-  d <- freq_table(x, reportvar)
+  #
+  if (update_NAs || update_values) {
+    d <- freq_table(jn, reportvar)
+  } else {
+    d <- report_from_attr(jn, y, reportvar)
+  }
+  # remove collapse::join attributes
+  attr(jn, "join.match") <- NULL
+
   rlang::env_poke(.joynenv, "freq_joyn", d)
 
   # Report var
   if (dropreport) {
-    get_vars(x, reportvar) <- NULL
+    get_vars(jn, reportvar) <- NULL
   }
 
   # store timing
@@ -517,15 +535,11 @@ joyn <- function(x,
     is executed in  ", round(time_taken_joyn, 6)))
 
   # return messages
-  if (verbose == TRUE) {
-    cli::cli_h2("JOYn Report")
-    joyn_report()
-    cli::cli_rule(right = "End of {.field JOYn} report")
-    joyn_msg(msg_type)
-  }
+  joyn_report(verbose = verbose)
+  if (verbose == TRUE) joyn_msg(msg_type)
 
-  setattr(x, "class", class_x)
+  setattr(jn, "class", class_x)
 
-  x
+  jn
 
 }