+++ NEWS.md
@@ -7,11 +7,15 @@
 
 * Add information about duplicated obs in `by` variable when match type is `1` rathern than `m`. 
 
-* improve ineffciencies in deep copies with `m:m` joins
+* improve inefficiencies in deep copies with `m:m` joins
 
 * Replace `m:m` joins from `data.table::merge.data.table` to `collapse::join`. Thanks to @SebKrantz for the suggestion (#58).
 
 * Add information about duplicated obs in `by` variable when match type is `1` rather than `m`.
+
+* Internal: improve storing of joyn messages.
+
+* Improve creation of reporting variable. Now, it is created in [collapse::join] rather than in `joyn` function. In addition, the reporting variable is created as factor to improve performance. Thanks to @SebKrantz for the suggestion (#58)
 
 ## breaking changes
 
@@ -19,6 +23,8 @@
 computational time that most of the time is not needed. 
 If the user wants to sort the data, they can use the `sort` argument, which triggers 
 the sorting mechanism of `collapse` package.
+
+* report variable (named ".join" by default) is now a factor instead of character. Yet, users can still use character if they want with the `reporttype = "character"`.
 
 # joyn 0.2.0